plot heatmap to visualize expression of provided key markers in each cell cluster or experimental group

plot heatmap to visualize expression of provided key markers in each cell cluster or experimental group

Usage

annotation_heatmap(data, gene, group)

Arguments

data

A Seurat object

gene

a list of genes for heatmap eg. gene<-c("Cd44","Cd69","Itgae","Ccr7","Il7r","Tbx21","Ifng","Il2","Gata3","Il4","Il5","Foxp3","Ctla4","Pdcd1","Tigit","Il2ra","Gzmb","Il10","Il17a","Cxcr5","Bcl6","Cd8a","Cd4")

group

the name of meta.data column within the Seurat object that indicates the cell cluster eg. group<-"seurat_clusters"

Value

A heatmap plot showing mean expression or percentage of cells that have no-zero expression of each gene in each cell cluster

Details

DETAILS

Examples

if (FALSE) {
if(interactive()){
 #EXAMPLE1
 annotation_heatmap(skin_tcell,gene = c("Cd44","Cd69"), group = "seurat_clusters")
 }
}